+++ Processors/ContainerProcessor.cs
@@ -251,11 +251,14 @@ sealed class ContainerProcessor : Processor
             var text = zdo.Vars.GetText();
             if (string.IsNullOrEmpty(text))
                 text = Config.Containers.ChestSignsDefaultText.Value;
+            var defaultColor = zdo.Vars.GetDefaultColor();
             ExtendedZDO sign;
             if (signOptions.HasFlag(ModConfig.ContainersConfig.SignOptions.Left))
             {
                 sign = PlacePiece(p + r * Vector3.right * signOffset.x, Prefabs.Sign, rot);
                 sign.Vars.SetText(text);
+                if (!string.IsNullOrEmpty(defaultColor))
+                    sign.Vars.SetDefaultColor(defaultColor);
                 signs.Add(sign);
                 _chestsBySigns.Add(sign, zdo);
             }
@@ -263,6 +266,8 @@ sealed class ContainerProcessor : Processor
             {
                 sign = PlacePiece(p + r * Vector3.left * signOffset.x, Prefabs.Sign, rot + 180);
                 sign.Vars.SetText(text);
+                if (!string.IsNullOrEmpty(defaultColor))
+                    sign.Vars.SetDefaultColor(defaultColor);
                 signs.Add(sign);
                 _chestsBySigns.Add(sign, zdo);
             }
@@ -270,6 +275,8 @@ sealed class ContainerProcessor : Processor
             {
                 sign = PlacePiece(p + r * Vector3.forward * signOffset.z, Prefabs.Sign, rot + 270);
                 sign.Vars.SetText(text);
+                if (!string.IsNullOrEmpty(defaultColor))
+                    sign.Vars.SetDefaultColor(defaultColor);
                 signs.Add(sign);
                 _chestsBySigns.Add(sign, zdo);
             }
@@ -277,6 +284,8 @@ sealed class ContainerProcessor : Processor
             {
                 sign = PlacePiece(p + r * Vector3.back * signOffset.z, Prefabs.Sign, rot + 90);
                 sign.Vars.SetText(text);
+                if (!string.IsNullOrEmpty(defaultColor))
+                    sign.Vars.SetDefaultColor(defaultColor);
                 signs.Add(sign);
                 _chestsBySigns.Add(sign, zdo);
             }